+++ src/main/java/com/uisrael/proyecto_clases/modelo/DAO/IRequerimientoDAO.java
@@ -21,7 +21,7 @@ public interface IRequerimientoDAO {
 
     public List<Requerimiento> findAll();
 
-    public boolean eliminar(int code);
+    public boolean eliminar(Long code);
 
     public boolean actualizar(Requerimiento objeto);
